pyproject.toml
[project]
name = 'syrupy'
license = 'MIT'
dynamic = [ "version", "classifiers" ]
description = 'Pytest Snapshot Test Utility'
readme = 'README.md'
requires-python = ">=3.10"
authors = [{name = 'Noah Ulster'}]
dependencies = [
  "pytest (>= 8.0.0)"
]

[project.urls]
repository = 'https://github.com/syrupy-project/syrupy'

[tool.poetry]
version = "5.0.0"
packages = [{ include = 'syrupy', from = 'src' }]
classifiers = [
    'Development Status :: 5 - Production/Stable',
    'Framework :: Pytest',
    'Intended Audience :: Developers',
    'Operating System :: OS Independent',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Programming Language :: Python :: 3.12',
    'Programming Language :: Python :: 3.13',
    'Topic :: Software Development :: Libraries',
    'Topic :: Software Development :: Testing',
    'Topic :: Utilities',
]
include = ['src/syrupy/**/*']

[tool.poetry.plugins.pytest11]
syrupy = 'syrupy'

[tool.poetry.group.test.dependencies]
invoke = "^2.2.0"
coverage = { version = "^7.6.12", extras = ['toml'] }
pytest-benchmark = "^5.1.0"
pytest-xdist = "^3.6.1"

[tool.poetry.group.dev.dependencies]
mypy = "^1.15.0"
py-githooks = '^1.1.1'
twine = '^6.1.0'
semver = '^3.0.4'
setuptools-scm = '^9.0.0'
debugpy = '^1.8.12'
ruff = '>=0.9.6'
pydantic = "^2.12.5"
attrs = "^23.2.0"

[tool.pytest.ini_options]
addopts = '-p syrupy -p pytester -p no:legacypath --doctest-modules'
testpaths = ['tests']
xfail_strict = true

[tool.coverage.run]
source = ['./src']

[tool.coverage.report]
exclude_lines = ['pragma: no-cover', 'if TYPE_CHECKING:', '@abstractmethod']

[tool.ruff]
line-length = 88
unsafe-fixes = true

[tool.ruff.lint]
extend-select = [
  "A", # flake8-builtins
  "B", # flake8-bugbear
  "C4", # flake8-comprehensions
  "I", # isort
  "UP", # pyupgrade
]

[build-system]
requires = ['poetry-core>=2.2.1']
build-backend = 'poetry.core.masonry.api'
